Optimize `arr = []` queries
We can use the same doc-value count as approximation before running the
generic function filter.

This helps a lot for tables where there are few empty arrays:

    Q: select count(*) from t_arrays_biased_5_elements where xs = []
    C: 1
    | Version |         Mean ±    Stdev |        Min |     Median |         Q3 |        Max |
    |   V1    |     1183.464 ±  941.087 |    519.783 |   1125.056 |   1426.058 |   6537.926 |
    |   V2    |      166.165 ±  656.880 |     26.856 |     36.494 |     46.460 |   7236.325 |
    ├---------┴-------------------------┴------------┴------------┴------------┴------------┘
    |               - 150.75%                           - 187.43%
    There is a 100.00% probability that the observed difference is not random, and the best estimate of that difference is 150.75%
    The test has statistical significance

And adds a slight cost to tables where the majority of the arrays is
empty:

    Q: select count(*) from t_arrays_biased_empty where xs = []
    C: 1
    | Version |         Mean ±    Stdev |        Min |     Median |         Q3 |        Max |
    |   V1    |      429.168 ±  400.471 |     84.141 |    246.755 |    831.188 |   2981.456 |
    |   V2    |      457.291 ±  407.171 |     85.589 |    403.508 |    833.148 |   3217.098 |
    ├---------┴-------------------------┴------------┴------------┴------------┴------------┘
    |               +   6.34%                           +  48.21%
    There is a 51.34% probability that the observed difference is not random, and the best estimate of that difference is 6.34%
    The test has no statistical significance
